requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
numpy>=1.24.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
